bert/pyproject.toml
[tool.poetry]
name = "ntng-bert"
version = "0.1.0"
description = ""
authors = ["Jean Kaddour"]
readme = "README.md"
packages = [
    { include = "cramming" },
    { include = "efficient_training" },
    { include = "rst" },
]

[tool.poetry.dependencies]
python = "~3.10"
hydra-core = ">=1.1"
torch = "~1.13+cu117"
datasets = "^2.13.1"
tokenizers = "^0.13.3"
transformers = "^4.30.2"
evaluate = "^0.4.0"
psutil = "^5.9.5"
einops = "^0.6.1"
zstandard = "^0.21.0"
wandb = "^0.15.5"
scipy = "^1.11.1"
scikit-learn = "^1.3.0"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
mypy = "^1.4.1"
isort = "^5.12.0"
pytest = "^7.4.0"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[[tool.poetry.source]]
name = "pytorch_cuda_117"
url = "https://download.pytorch.org/whl/cu117"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 140

[tool.isort]
known_third_party = ["wandb"]
profile = "black"

[tool.pytest.ini_options]
pythonpath = ["."]
